Health Bar

+++ Tower Defense/Assets/Scripts/Data/RawEnemyData.cs
@@ -3,8 +3,8 @@
     public int enemyID;
     public string enemyName;
     public string enemyType;
-    public int enemyHP;
-    public int enemyMaxHP;
+    public float enemyHP;
+    public float enemyMaxHP;
     public int enemyDamage;
     public int enemyDefense;
     public float enemyMoveSpd;
